John Loke's patches regarding receipt sources and invoice quantities, and adding recommendation for Math::BigInt::GMP to Makefile
git-svn-id: svn://svn.code.sf.net/p/ledger-smb/code/trunk@3286 4979c152-3d1c-0410-bac9-87ea11338e46

+++ sql/Pg-database.sql
@@ -885,7 +885,7 @@ CREATE TABLE invoice (
   trans_id int,
   parts_id int,
   description text,
-  qty integer,
+  qty NUMERIC,
   allocated integer,
   sellprice NUMERIC,
   precision int,
